fix delete user

+++ routes/web.php
@@ -25,12 +25,16 @@ Route::group(['namespace' => 'Auth', 'middleware' => 'auth'], function () {
     Route::get('admin/teams/{team}/edit', '\App\Http\Controllers\Admin\Team\EditController')->name('admin.team.edit');
     Route::patch('admin/teams/{team}', '\App\Http\Controllers\Admin\Team\UpdateController')->name('admin.team.update');
     Route::delete('admin/teams/{team}', '\App\Http\Controllers\Admin\Team\DestroyController')->name('admin.team.delete');
+
     Route::get('admin/users', '\App\Http\Controllers\Admin\User\IndexController')->name('admin.user.index');
     Route::get('admin/users/create', '\App\Http\Controllers\Admin\User\CreateController')->name('admin.user.create');
     Route::post('admin/users', '\App\Http\Controllers\Admin\User\StoreController')->name('admin.user.store');
     Route::get('admin/users/{user}/edit', '\App\Http\Controllers\Admin\User\EditController')->name('admin.user.edit');
     Route::patch('admin/users/{user}', '\App\Http\Controllers\Admin\User\UpdateController')->name('admin.user.update');
-    Route::delete('', '\App\Http\Controllers\Admin\User\DestroyController')->name('admin.user.delete');
+//    Route::delete('', '\App\Http\Controllers\Admin\User\DestroyController')->name('admin.user.delete');
+    Route::delete('/admin/user/{user}', '\App\Http\Controllers\Admin\User\DestroyController')->name('admin.user.delete');
+
+
     Route::get('admin/setting-prices', [\App\Http\Controllers\Admin\SettingPricesController::class, 'index'])->name('admin.settingPrices.index');
 
     Route::get('/', [\App\Http\Controllers\DashboardController::class, 'index'])->name('dashboard');
